change handling the modified flag of a view to use a signed counter
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@29629 cc8e0003-5e82-43e2-9771-63201f7d8d9c

+++ src/TortoiseMerge/Undo.h
@@ -116,9 +116,9 @@ protected:
     std::list<POINT>                     m_redoCaretPoints;
     std::list<std::list<int>::size_type> m_redoGroups;
 
-    size_t                               m_savedStateIndexLeft;
-    size_t                               m_savedStateIndexRight;
-    size_t                               m_savedStateIndexBottom;
+    __int64                              m_savedStateIndexLeft;
+    __int64                              m_savedStateIndexRight;
+    __int64                              m_savedStateIndexBottom;
 
 private:
     CUndo();
